backend/requirements.txt
# File: backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
supabase==2.5.0
pydantic==2.7.1
pydantic-settings==2.2.1
google-generativeai==0.5.4
httpx==0.27.0